Revert handler.py to pre-raw-detector-view and extract raw-view
